IGNITE-17676 Schema updates.
